Extract Cache to types/ directory
