+++ packages/dictionaries/src/lib/LayeredCache.ts
@@ -1,7 +1,8 @@
 import { Trie } from '@kamilmielnik/trie';
 import { Locale } from '@scrabble-solver/types';
 
-import Cache from './Cache';
+import { Cache } from '../types';
+
 import DiskCache from './DiskCache';
 import MemoryCache from './MemoryCache';
