+++ packages/dictionaries/src/lib/MemoryCache.ts
@@ -2,8 +2,7 @@ import { Trie } from '@kamilmielnik/trie';
 import { Locale } from '@scrabble-solver/types';
 
 import { CACHE_STALE_THRESHOLD } from '../constants';
-
-import Cache from './Cache';
+import { Cache } from '../types';
 
 class MemoryCache implements Cache<Locale, Trie> {
   private readonly cache: Partial<Record<Locale, Trie>> = {};
