+++ packages/dictionaries/src/lib/index.ts
@@ -1,4 +1,3 @@
-export { default as Cache } from './Cache';
 export { default as createDownloadDictionaryProxies } from './createDownloadDictionaryProxies';
 export { default as createDownloadDictionaryProxy } from './createDownloadDictionaryProxy';
 export { default as DiskCache } from './DiskCache';
