+++ packages/dictionaries/src/types/index.ts
@@ -1,0 +1,1 @@
+export { default as Cache } from './Cache';
